Moving tokens.ts into src/ and out of src/global.
