+++ packages/react-theme/src/global/index.ts
@@ -2,4 +2,3 @@ export * from './colors';
 export * from './fonts';
 export * from './strokeWidths';
 export * from './borderRadius';
-export * from './tokens';
